+++ ddbmock/database/key.py
@@ -5,13 +5,27 @@
 from ddbmock.errors import ValidationException
 
 class Key(object):
+    """
+    Abstraction layer over DynamoDB Keys in :py:class:`ddbmock.databas.item.Item`
+    """
     def __init__(self, name, typename):
+        """
+        High level Python constructor
+
+        :param name: Valid key name. No further checks are performed.
+        :param typename: Valid key typename. No further checks are performed.
+        """
         self.name = name
         self.typename = typename
 
     def read(self, key):
-        """Parse a key as specified by DynamoDB API and return its value as long as
-            its typename matches self.typename
+        """
+        Parse a key as specified by DynamoDB API and return its value as long as
+        its typename matches :py:attr:`typename`
+
+        :param key: Raw DynamoDB request key.
+
+        :return: the value of the key
         """
         typename, value = key.items()[0]
         if self.typename != typename:
@@ -21,7 +35,10 @@ class Key(object):
         return value
 
     def to_dict(self):
-        """Return the a dict form of the key, suitable for DynamoDb API
+        """
+        Return the key as a Python dict.
+
+        :return: Serialized version of the key definition metadata compatible with DynamoDB API syntax.
         """
         return {
             "AttributeName": self.name,
@@ -30,7 +47,18 @@ class Key(object):
 
     @classmethod
     def from_dict(cls, data):
+        """
+        Alternate constructor which deciphers raw DynamoDB request data before
+        ultimately calling regular ``__init__`` method.
+
+        See :py:meth:`__init__` for more insight.
+
+        :param data: raw DynamoDB request data.
+
+        :return: fully initialized :py:class:`Key` instance
+        """
         return cls(data[u'AttributeName'], data[u'AttributeType'])
 
 class PrimaryKey(Key):
+    """Special marker to provide distinction between regulat Keys and PrimaryKey"""
     pass